feat: Add API Gateway security filter
The Route Validator contains a Predicate which decides if a route should
be secured or not. After that, the AuthenticationFilter intercepts given
routes and applies it filtering method by which he checks for the
Authorization header for the JWT token which he validates by calling the
auth service and it's validateToken endpoint. The
ApiGatewayConfiguration for now only applies it to the currency service.