feat: Add user service
The user service replaces the old repository and functionality inside
the authentication filter and it now communicates with the
authentication filter via open endpoints. It has a limited amount of
endpoints for the security of the users credentials. The saveUser
endpoints saves a user to the database if it doesn't already exist. If
it does it will return null. There is possibly an unnecessary exception
throwing which will probably be patched later. The findByUsername
endpoint returns a User with the given username if he exists. Otherwise,
it will return null.