feat: Add JWT service
The whole JWT token functionality has been removed from the
authentication service to its own service. Methods for claim extraction
have been added for the purpose of adding the logged user's username to
the request headers by the AuthenticationFilter. The behaviour of the
API has been about the same. This allows for more modularity and
independece down the line.