feat: Add getMonthStats endpoint
The method now returns a new object from the StatsResults class. The
class contains a Map which is parametrized by String and Double. It also
contains the grouping element and the current date and time. First, the
service gets all Expenses from a username, which the service forwards
through as a request header to the expense service using OpenFeign, and
then it group by currencies and calculates the sum of the amounts. The
resulting map is added to the new StatsResults object and is sent back.